Moved most of stnode code into lib.
