+++ components/bin/src/lib.rs
@@ -13,3 +13,4 @@
 
 pub mod stindex;
 pub mod stmgr;
+pub mod stnode;
